+++ src/include/generator_v7.hpp
@@ -1,0 +1,21 @@
+#pragma once
+
+#include "generator.hpp"
+#include "mutator.hpp"
+
+/*
+ * Generator that uses pattern files. A pattern is a program with annotation
+ * that indicate positions where mutations should be applied.
+ */
+class GeneratorV7 : public Generator {
+ public:
+  GeneratorV7(const Config &config, const Stats &stats);
+
+  virtual Program generateProgram() override;
+
+  virtual std::pair<Operation, double> generateOperation() override;
+
+ private:
+  std::vector<Program> patterns;
+  Mutator mutator;
+};